# Conflicts:
#	lib/app/modules/single_shop/product_list/data/remote/models/product_response_model.dart
#	lib/app/modules/single_shop/sell/presentation/pages/quick_sell_front.dart

+++ lib/app/modules/single_shop/sell/presentation/pages/sell_page.dart
@@ -1648,7 +1648,7 @@ List<Widget> getTabBarViews(
                                       Product product = Product(
                                           name: controller
                                               .customProductName.value,
-                                          sellingPrice: double.parse(controller
+                                          sellingPrice: int.parse(controller
                                               .customProductPrice.value),
                                           stock: 1);
